[NMPC][Refactor]: make a base_mpc_solver class and try to move most implementations from nmpc solver to this base class.
